pyproject.toml
[project]
name = "bps-stat-agent"
version = "1.0.0"
description = "BPS Indonesia Statistical Data Agent & Academic Research AI Agent"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Julio Christian Wadu Doko", email = "juliochwd@gmail.com"}
]
license = {text = "MIT"}
keywords = ["bps", "statistics", "indonesia", "ntt", "agent", "mcp", "badan-pusat-statistik"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "mcp>=1.23.0",
    "requests>=2.33.0",
    "tiktoken>=0.5.0",
    "prompt-toolkit>=3.0.0",
    "anthropic>=0.39.0",
    "openai>=1.57.4",
    "agent-client-protocol>=0.6.0",
    "playwright>=1.40.0",
    # Transitive dep pins for security (CVE fixes)
    "urllib3>=2.6.3",           # CVE-2026-21441, CVE-2025-66471, CVE-2025-66418
    "python-multipart>=0.0.26", # CVE-2026-24486, CVE-2026-40347
    "starlette>=0.49.0",        # CVE-2025-62727
    "python-dotenv>=1.2.2",     # Symlink following in set_key
    "Pygments>=2.20.0",         # ReDoS via GUID matching
]

[project.scripts]
bps-stat-agent = "mini_agent.cli:main"
bpsagent = "mini_agent.cli:main"
bps-stat-agent-acp = "mini_agent.acp.server:main"
bps-mcp-server = "mini_agent.bps_mcp_server:main"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.2.0",
]
metrics = [
    "prometheus-client>=0.20.0",
]
tracing = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
observability = [
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
research-core = [
    "statsmodels>=0.14.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "networkx>=3.2.0",
    "bibtexparser>=1.4.0",
]
research-analysis = [
    "bps-stat-agent[research-core]",
    "linearmodels>=6.0",
    "arch>=7.0",
    "pingouin>=0.5.4",
]
research-advanced-stats = [
    "bps-stat-agent[research-analysis]",
    "pymc>=5.0",
    "bambi>=0.14",
    "arviz>=0.18",
    "dowhy>=0.11",
    "lifelines>=0.29",
]
research-document = [
    "markitdown[all]>=0.1.5",
    "chonkie>=0.4",
]
research-rag = [
    "sentence-transformers>=3.0",
    "lancedb>=0.9",
    "lightrag-hku>=1.0",
]
research-writing = [
    "pylatex>=1.4.2",
    "python-docx>=1.2",
    "bibtexparser>=1.4",
    "habanero>=2.3",
    "textstat>=0.7.0",
    "proselint>=0.14.0",
]
research-quality = [
    "language-tool-python>=2.8",
    "pandera>=0.20",
]
research-sandbox = [
    "docker>=7.0",
    "RestrictedPython>=7.0",
]
research-nlp = [
    "spacy>=3.7",
    "scispacy>=0.5",
]
research-llm = [
    "litellm>=1.40",
    "dspy>=2.0",
]
research-all = [
    "bps-stat-agent[research-core]",
    "bps-stat-agent[research-analysis]",
    "bps-stat-agent[research-advanced-stats]",
    "bps-stat-agent[research-document]",
    "bps-stat-agent[research-rag]",
    "bps-stat-agent[research-writing]",
    "bps-stat-agent[research-quality]",
    "bps-stat-agent[research-sandbox]",
    "bps-stat-agent[research-nlp]",
    "bps-stat-agent[research-llm]",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
mini_agent = ["skills/**/*", "config/*"]

[tool.setuptools.packages.find]
include = ["mini_agent*"]
exclude = ["tests*", "plan*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
cache_dir = "workspace/.pytest_cache"
asyncio_mode = "auto"
markers = [
    "live: tests that hit real upstream BPS services",
]
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning",
    "ignore::DeprecationWarning:mcp.*",
]

[tool.pylint.messages_control]
# Disable warnings for Tool.execute method signature differences
# We intentionally use explicit parameters in subclasses for better type hints
disable = [
    "arguments-differ",  # Allow subclasses to have different parameter signatures
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.8.0",
    "pip-audit>=2.7.0",
]
